搜索记录 异步调用 接口Bug修改

+++ src/main/java/com/itswpu/huanswpu/service/UserService.java
@@ -13,7 +13,7 @@ public interface UserService extends IService<User> {
      * @param keyword
      * @param userId
      */
-    public void insert(String keyword,Integer userId);
+    public void insert(String keyword,Long userId);
 
     /**
      查询搜索历史
